+++ modules/account/variables.tf
@@ -6,15 +6,18 @@ variable "profile" {
   description = "The profile name as set in the shared credentials file. If not set, it will be sourced from the ALICLOUD_PROFILE environment variable."
   default     = ""
 }
+
 variable "shared_credentials_file" {
   description = "This is the path to the shared credentials file. If this is not set and a profile is specified, $HOME/.aliyun/config.json will be used."
   default     = ""
 }
+
 variable "skip_region_validation" {
   description = "Skip static validation of region ID. Used by users of alternative AlibabaCloud-like APIs or users w/ access to regions that are not public (yet)."
   type        = bool
   default     = false
 }
+
 variable "region" {
   description = "The region used to launch this module resources."
   default     = ""
@@ -25,7 +28,7 @@ variable "region" {
 ################
 
 variable "create_account" {
-  description = "Whether to create a new account. If true, the `accounts` should not be empty."
+  description = "Whether to create a new account. If true, the 'accounts' should not be empty."
   type        = bool
   default     = true
 }